phase_5_api/requirements.txt
fastapi
uvicorn
pydantic
pytest
httpx
google-generativeai
python-dotenv
